Split button combos into new file
