+++ fw/combo.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include "config.h"
+
+void process_combos(config* current_config,
+                    timer* combo_timer,
+                    timer* combo_lights_timer);
